chore: create funnels.py, move funnel hints into separate method
