+++ analytics_mcp/server.py
@@ -25,6 +25,7 @@ from analytics_mcp.coordinator import mcp
 from analytics_mcp.tools.admin import info  # noqa: F401
 from analytics_mcp.tools.reporting import realtime  # noqa: F401
 from analytics_mcp.tools.reporting import core  # noqa: F401
+from analytics_mcp.tools.reporting import funnel  # noqa: F401
 
 
 def run_server() -> None:
